calculates TFactS

This function calculates TFactS

calculateTFactS(
  DEGs,
  catalog,
  TFs,
  all.targets,
  Q.value = FALSE,
  lambda1 = seq(0.05, 0.5, 0.01),
  lambda2 = 0.05,
  nRep = 100,
  TF.col = "TF..OFFICIAL_TF_CODING_GENE_NAME."
)

Arguments

DEGs

a character vector of DEGs (differentially expressed genes)

catalog

a data frame of TFactS catalog (ver. 2)

TFs

a character vector of transcription factor

all.targets

a character vector of all target genes

Q.value

logical. If it is TRUE, Q.value calculation is performed by Storey method.

lambda1

a vector of the lambda values utilized to obtain pi0.lambda

lambda2

a user-specified threshold of E-value (default: 0.05)

nRep

number of random selections (default: 100)

TF.col

the name of the column that contains the TF names

Value

data.frame

References

Essaghir A et al. Nucleic Acids Res. 2010 Jun;38(11):e120.

Author

Atsushi Fukushima

Examples

data(DEGs)
data(catalog)

tftg <- extractTFTG(DEGs, catalog)
TFs <- tftg$TFs
all.targets <- tftg$all.targets

res <- calculateTFactS(DEGs, catalog, TFs, all.targets)